exception creados

+++ AlertaHospital/src/main/java/com/example/AlertaHospital/service/alertaService.java
@@ -1,5 +1,6 @@
 package com.example.AlertaHospital.service;
 
+import com.example.AlertaHospital.exception.ResourceNotFoundException;
 import com.example.AlertaHospital.dto.alertaDTO;
 import com.example.AlertaHospital.model.alertaModel;
 import com.example.AlertaHospital.repository.alertaRepository;
@@ -15,7 +16,6 @@ public class alertaService {
     @Autowired
     private alertaRepository repository;
 
-
     private alertaDTO mapToDTO(alertaModel model) {
         return alertaDTO.builder()
                 .idAlerta(model.getIdAlerta())
@@ -25,7 +25,6 @@ public class alertaService {
                 .estado(model.getEstado())
                 .build();
     }
-
 
     private alertaModel mapToEntity(alertaDTO dto) {
         return alertaModel.builder()
@@ -37,20 +36,17 @@ public class alertaService {
                 .build();
     }
 
-
     public List<alertaDTO> getAllAlertas() {
         return repository.findAll().stream()
                 .map(this::mapToDTO)
                 .collect(Collectors.toList());
     }
 
-
     public alertaDTO getAlertaById(Long id) {
         alertaModel alerta = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Alerta no encontrada con ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Alerta no encontrada con ID: " + id));
         return mapToDTO(alerta);
     }
-
 
     public alertaDTO createAlerta(alertaDTO dto) {
         alertaModel alerta = mapToEntity(dto);
@@ -58,10 +54,9 @@ public class alertaService {
         return mapToDTO(nuevaAlerta);
     }
 
-
     public alertaDTO updateAlerta(Long id, alertaDTO dto) {
         alertaModel alerta = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Alerta no encontrada con ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Alerta no encontrada con ID: " + id));
 
         alerta.setTipo(dto.getTipo());
         alerta.setMensaje(dto.getMensaje());
@@ -72,8 +67,10 @@ public class alertaService {
         return mapToDTO(alertaActualizada);
     }
 
-
     public void deleteAlerta(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Alerta no encontrada con ID: " + id);
+        }
         repository.deleteById(id);
     }
 }
